refactor files into utils folder to avoid loading/init error
