+++ Track/Data/Transforms-multiple-columns.csv
@@ -1,0 +1,72 @@
+X,Y,Z,something_else,random,Y1,Z1,X1
+-3.758735288,-9.218964113,7.535308611,0.698350092,0.529092495,-9.218964113,7.535308611,-3.758735288
+-3.670512893,-9.218964113,6.171260311,0.556587314,0.808811385,-9.218964113,6.171260311,-3.670512893
+-3.670512893,-10.1624006,6.064666992,0.546810745,0.118189661,-10.1624006,6.064666992,-3.670512893
+-3.847001835,-10.1624006,4.713925897,0.539816458,0.375050116,-10.1624006,4.713925897,-3.847001835
+-3.847001835,-10.18379727,5.038935286,0.06141077,0.681452305,-10.18379727,5.038935286,-3.847001835
+-4.424329736,-10.18379727,3.788098755,0.251801951,0.266377362,-10.18379727,3.788098755,-4.424329736
+-4.424329736,-10.56460581,4.591186414,0.915070578,0.913658222,-10.56460581,4.591186414,-4.424329736
+-4.635133717,-10.56460581,3.316021327,0.254822239,0.883658553,-10.56460581,3.316021327,-4.635133717
+-4.635133717,-10.03276018,4.387041828,0.530083159,0.029715113,-10.03276018,4.387041828,-4.635133717
+-4.667831165,-10.03276018,3.205307475,0.882620026,0.941438454,-10.03276018,3.205307475,-4.667831165
+-4.667831165,-9.887201242,5.005008254,0.621115123,0.916546762,-9.887201242,5.005008254,-4.667831165
+-3.66578249,-9.887201242,5.262222132,0.972143021,0.636501992,-9.887201242,5.262222132,-3.66578249
+-3.66578249,-9.537811807,6.97284004,0.817567556,0.608638052,-9.537811807,6.97284004,-3.66578249
+-1.72096438,-9.537811807,8.554314594,0.556719329,0.495488979,-9.537811807,8.554314594,-1.72096438
+-1.72096438,-8.276002601,9.539078857,0.491649595,0.148075211,-8.276002601,9.539078857,-1.72096438
+-1.299154625,-8.276002601,9.834385343,0.520537814,0.978751124,-8.276002601,9.834385343,-1.299154625
+-1.299154625,-8.036132831,10.22337296,0.5419711,0.995850305,-8.036132831,10.22337296,-1.299154625
+-3.006020218,-8.036132831,8.634480796,0.894516125,0.14986301,-8.036132831,8.634480796,-3.006020218
+-3.006020218,-8.967566049,8.748662331,0.564300912,0.00164453,-8.967566049,8.748662331,-3.006020218
+-3.600435534,-8.967566049,6.640702554,0.629734668,0.99346659,-8.967566049,6.640702554,-3.600435534
+-3.600435534,-10.02609017,6.400379188,0.335511206,0.03468567,-10.02609017,6.400379188,-3.600435534
+-3.729867756,-10.02609017,5.16128228,0.114042778,0.148092203,-10.02609017,5.16128228,-3.729867756
+-3.729867756,-10.17180535,5.57434365,0.147719832,0.841998523,-10.17180535,5.57434365,-3.729867756
+-3.570926883,-10.17180535,5.687644524,0.293062841,0.93394799,-10.17180535,5.687644524,-3.570926883
+-3.570926883,-9.702920089,7.323810753,0.658498861,0.683592121,-9.702920089,7.323810753,-3.570926883
+-2.456734537,-9.702920089,8.849544283,0.603827666,0.597556443,-9.702920089,8.849544283,-2.456734537
+-2.456734537,-8.241878968,10.52553786,0.876115519,0.12470944,-8.241878968,10.52553786,-2.456734537
+-0.480356784,-8.241878968,10.91979952,0.428517341,0.229658331,-8.241878968,10.91979952,-0.480356784
+-0.480356784,-6.994417409,12.1795903,0.904427373,0.142665598,-6.994417409,12.1795903,-0.480356784
+-1.709030604,-6.994417409,10.47676466,0.624025474,0.227796232,-6.994417409,10.47676466,-1.709030604
+-1.709030604,-7.59384935,11.54739901,0.883484124,0.066546316,-7.59384935,11.54739901,-1.709030604
+-3.075296368,-7.59384935,9.085298703,0.196609958,0.58084397,-7.59384935,9.085298703,-3.075296368
+-3.075296368,-8.494823301,9.66166908,0.583689328,0.933372898,-8.494823301,9.66166908,-3.075296368
+-3.622081698,-8.494823301,7.341445152,0.212995755,0.669316451,-8.494823301,7.341445152,-3.622081698
+-3.622081698,-9.537036266,7.285647866,0.571889309,0.980270385,-9.537036266,7.285647866,-3.622081698
+-3.807722974,-9.537036266,5.711744493,0.997732176,0.813776486,-9.537036266,5.711744493,-3.807722974
+-3.807722974,-10.13479764,5.740263253,0.333148681,0.32000853,-10.13479764,5.740263253,-3.807722974
+-3.742005965,-10.13479764,4.704025199,0.046752595,0.472809363,-10.13479764,4.704025199,-3.742005965
+-3.742005965,-10.09993939,5.45960854,0.166754138,0.786627427,-10.09993939,5.45960854,-3.742005965
+-3.700780187,-10.09993939,5.577864991,0.323373937,0.255715644,-10.09993939,5.577864991,-3.700780187
+-3.700780187,-9.69880476,7.273258506,0.951119666,0.065652461,-9.69880476,7.273258506,-3.700780187
+-2.694631177,-9.69880476,8.553740616,0.21491236,0.307350711,-9.69880476,8.553740616,-2.694631177
+-2.694631177,-8.169201783,10.47978624,0.160011012,0.379977644,-8.169201783,10.47978624,-2.694631177
+-1.374563464,-8.169201783,10.45638433,0.836779902,0.643655282,-8.169201783,10.45638433,-1.374563464
+-1.374563464,-7.008312284,12.16533933,0.327374355,0.534297036,-7.008312284,12.16533933,-1.374563464
+-2.111060794,-7.008312284,10.22548535,0.485580712,0.230394314,-7.008312284,10.22548535,-2.111060794
+-2.111060794,-7.693453086,11.52626413,0.916535955,0.208072217,-7.693453086,11.52626413,-2.111060794
+-3.384413109,-7.693453086,9.038787512,0.903210709,0.49959766,-7.693453086,9.038787512,-3.384413109
+-3.384413109,-8.612755596,9.675737514,0.000518859,0.549742168,-8.612755596,9.675737514,-3.384413109
+-3.759871487,-8.612755596,7.249070456,0.246849505,0.562040128,-8.612755596,7.249070456,-3.759871487
+-3.759871487,-9.620274124,7.260049896,0.595411613,0.588889333,-9.620274124,7.260049896,-3.759871487
+-3.793720414,-9.620274124,5.789090506,0.881872863,0.085548627,-9.620274124,5.789090506,-3.793720414
+-3.793720414,-10.06118707,5.971900364,0.970208531,0.129881656,-10.06118707,5.971900364,-3.793720414
+-3.869556482,-10.06118707,4.967442312,0.495251743,0.806764161,-10.06118707,4.967442312,-3.869556482
+-3.869556482,-10.14611497,5.691884998,0.93487481,0.755446469,-10.14611497,5.691884998,-3.869556482
+-3.857025818,-10.14611497,5.439753451,0.777113283,0.094500823,-10.14611497,5.439753451,-3.857025818
+-3.857025818,-9.757940571,6.925915234,0.855774331,0.537401491,-9.757940571,6.925915234,-3.857025818
+-3.393010444,-9.757940571,7.424994156,0.086615022,0.64131694,-9.757940571,7.424994156,-3.393010444
+-3.393010444,-8.617292754,9.526228901,0.033826771,0.830309488,-8.617292754,9.526228901,-3.393010444
+-2.137178021,-8.617292754,9.648241467,0.740123656,0.107023016,-8.617292754,9.648241467,-2.137178021
+-2.137178021,-7.57405653,11.13000153,0.924472667,0.750416896,-7.57405653,11.13000153,-2.137178021
+-2.065086624,-7.57405653,9.748393148,0.167206816,0.334880229,-7.57405653,9.748393148,-2.065086624
+-2.065086624,-7.68218,10.73118453,0.04943115,0.229244094,-7.68218,10.73118453,-2.065086624
+-3.367688135,-7.68218,8.405566,0.234584321,0.145013719,-7.68218,8.405566,-3.367688135
+-3.367688135,-8.662943936,9.268749462,0.011149871,0.363565479,-8.662943936,9.268749462,-3.367688135
+-3.748341132,-8.662943936,7.047347184,0.592878649,0.12924957,-8.662943936,7.047347184,-3.748341132
+-3.748341132,-9.620587868,7.202484133,0.453898294,0.742109325,-9.620587868,7.202484133,-3.748341132
+-3.722858725,-9.620587868,5.824618003,0.010356639,0.545725634,-9.620587868,5.824618003,-3.722858725
+-3.722858725,-10.07234733,5.919251503,0.10649324,0.927177344,-10.07234733,5.919251503,-3.722858725
+-3.758735288,-10.07234733,4.860097248,0.906540173,0.290188056,-10.07234733,4.860097248,-3.758735288
+-3.758735288,-10.66501859,4.860097248,0.874756176,0.436063003,-10.66501859,4.860097248,-3.758735288
